requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.12.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
async-lru==2.0.5
asyncpg==0.30.0
attrs==25.3.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
ConfigArgParse==1.7
crashtest==0.4.1
cryptography==44.0.2
deprecation==2.1.0
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
dnspython==2.7.0
dulwich==0.22.8
ecdsa==0.19.1
email-validator==2.0.0
Faker==37.0.0
fastapi==0.103.1
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
frozenlist==1.6.0
gevent==24.11.1
geventhttpclient==2.3.3
gotrue==2.12.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
identify==2.6.9
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.6.0
locust==2.33.2
Mako==1.3.10
MarkupSafe==3.0.2
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.4.3
nodeenv==1.9.1
packaging==24.2
passlib==1.7.4
pbs-installer==2025.3.17
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.7
pluggy==1.6.0
plyer==2.1.0
poetry==2.1.2
poetry-core==2.1.2
postgrest==1.0.2
pre_commit==4.2.0
propcache==0.3.1
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.7
pyasn1==0.4.8
pycairo==1.28.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyGObject==3.52.3
PyJWT==2.10.1
pyproject_hooks==1.2.0
pystray==0.19.5
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.6
python-xlib==0.33
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
realtime==2.4.3
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
SecretStorage==3.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.20
sqlparse==0.5.3
starlette==0.27.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
tomlkit==0.13.2
trove-classifiers==2025.3.19.19
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.23.2
virtualenv==20.30.0
websockets==14.2
Werkzeug==3.1.3
yarl==1.20.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
